Fix dryrun output for Ansible
Dryrun is mind to print Terraform and Ansible commands that a user can
directly run. The first Ansible command produced an invalid command due
to missing quotation. This commit is about fixing this issue. The
approach is to change the internal rappresentation for the commands,
from a list of list of string (subprocess expect a list of strings for a
single command), to a list of strings: each command is just rappresented
with a single space separated string. This format can be directly used
for the dryrun mode. When not in dryrun mode, process_manager is using
shlex.split to produce the input subproces needs.